requirements.txt
yfinance==0.2.31
requests==2.31.0
beautifulsoup4==4.12.2
praw==7.7.1
transformers==4.36.2
torch>=2.0.0
scikit-learn==1.3.2
numpy==1.26.2
python-dotenv==1.0.0
nltk==3.8.1
